+++ Modules/Map/MapModule.lua
@@ -163,35 +163,33 @@ function MapModule:UpdateWorldMap()
     MapModule.WorldMapDataProvider:RefreshAllData()
 end
 
+local function BattlePetToggle_OnClick()
+    if ConfigModule.AceDB.profile.mapPinsToInclude == "T4NONE" then
+        ConfigModule.AceDB.profile.mapPinsToInclude = ConfigModule.AceDB.profile.mapPinsToIncludeOriginal
+    else
+        ConfigModule.AceDB.profile.mapPinsToIncludeOriginal = ConfigModule.AceDB.profile.mapPinsToInclude
+        ConfigModule.AceDB.profile.mapPinsToInclude = "T4NONE"
+    end
+
+    MapModule:UpdateWorldMap()
+end
+
 function MapModule:InitializeDropDown()
     AceHook:SecureHook(WorldMapFrame.overlayFrames[2], "InitializeDropDown", function()
         UIDropDownMenu_AddSeparator()
 
-        local header = 
-        {
-            isTitle = true,
-            notCheckable = true,
-            text = "Battle Pet Completionist",
-        }
+        local header = UIDropDownMenu_CreateInfo()
+        header.isTitle = true
+        header.notCheckable = true
+        header.text = "Battle Pet Completionist"
         UIDropDownMenu_AddButton(header)
 
-        local battlePetToggle =
-        {
-            isNotRadio = true,
-            keepShownOnClick = true,
-            text = "Battle Pets",
-            checked = ConfigModule.AceDB.profile.mapPinsToInclude ~= "T4NONE",
-            func = function()
-                if ConfigModule.AceDB.profile.mapPinsToInclude == "T4NONE" then
-                    ConfigModule.AceDB.profile.mapPinsToInclude = ConfigModule.AceDB.profile.mapPinsToIncludeOriginal
-                else
-                    ConfigModule.AceDB.profile.mapPinsToIncludeOriginal = ConfigModule.AceDB.profile.mapPinsToInclude
-                    ConfigModule.AceDB.profile.mapPinsToInclude = "T4NONE"
-                end
-
-                MapModule:UpdateWorldMap()
-            end
-        }
+        local battlePetToggle = UIDropDownMenu_CreateInfo()
+        battlePetToggle.isNotRadio = true
+        battlePetToggle.keepShownOnClick = true
+        battlePetToggle.text = "Battle Pets"
+        battlePetToggle.checked = ConfigModule.AceDB.profile.mapPinsToInclude ~= "T4NONE"
+        battlePetToggle.func = BattlePetToggle_OnClick
         UIDropDownMenu_AddButton(battlePetToggle)
     end)