+++ Changelog.md
@@ -1,3 +1,6 @@
+# NEXT
+* Changed the way UIDropDownMenu items are created
+
 # 10.1.0-20230617_1
 * Added possibility to toggle map pins from the world map overlay frame dropdown